refactor(*): restructure project files
