--- index.ts
@@ -1,5 +1,0 @@
-import { pocketbaseLoader } from "./src/pocketbase-loader";
-import type { PocketBaseLoaderOptions } from "./src/types/pocketbase-loader-options.type";
-
-export { pocketbaseLoader };
-export type { PocketBaseLoaderOptions };
