+++ package.json
@@ -7,10 +7,9 @@
   "homepage": "https://github.com/pawcoding/astro-loader-pocketbase",
   "type": "module",
   "exports": {
-    ".": "./index.ts"
+    ".": "./src/index.ts"
   },
   "files": [
-    "index.ts",
     "src"
   ],
   "scripts": {
