+++ src/index.ts
@@ -1,0 +1,5 @@
+import { pocketbaseLoader } from "./pocketbase-loader";
+import type { PocketBaseLoaderOptions } from "./types/pocketbase-loader-options.type";
+
+export { pocketbaseLoader };
+export type { PocketBaseLoaderOptions };
