+++ src/loader/load-entries.ts
@@ -1,6 +1,6 @@
 import type { LoaderContext } from "astro/loaders";
-import type { PocketBaseLoaderOptions } from "./types/pocketbase-loader-options.type";
-import { parseEntry } from "./utils/parse-entry";
+import type { PocketBaseLoaderOptions } from "../types/pocketbase-loader-options.type";
+import { parseEntry } from "./parse-entry";
 
 /**
  * Load (modified) entries from a PocketBase collection.
